move array stream xptr stuff to header
