+++ r/src/pointers.c
@@ -30,7 +30,7 @@ SEXP nanoarrow_c_allocate_schema(void) { return nanoarrow_schema_owning_xptr(); 
 
 SEXP nanoarrow_c_allocate_array(void) { return nanoarrow_array_owning_xptr(); }
 
-SEXP nanoarrow_c_allocate_array_stream(void) { return array_stream_owning_xptr(); }
+SEXP nanoarrow_c_allocate_array_stream(void) { return nanoarow_array_stream_owning_xptr(); }
 
 SEXP nanoarrow_c_pointer(SEXP obj_sexp) {
   if (TYPEOF(obj_sexp) == EXTPTRSXP) {
